+++ lib/runners/processor/cpplint.rb
@@ -26,17 +26,11 @@ module Runners
     CONFIG_FILE_NAME = "CPPLINT.cfg".freeze
 
     def setup
-      # TODO: When the notification period expires, comment out the below line (warn_recommended_config_file_release)
-      #       and uncomment the second line (deploy_recommended_config_file) to activate our default configuration file.
-      warn_recommended_config_file_release(CONFIG_FILE_NAME, "in mid October 2020")
-
-=begin
       if config_linter[:filter]
-        trace_writer.message "The `filter` option in #{config.path_name} is specified. The Sider's recommended ruleset is ignored."
+        trace_writer.message "The `filter` option in `#{config.path_name}` is specified. The Sider's recommended ruleset is ignored."
       else
         deploy_recommended_config_file(CONFIG_FILE_NAME)
       end
-=end
 
       yield
     end