+++ test/smokes/cpplint/expectations.rb
@@ -177,55 +177,46 @@ s.add_test(
 s.add_test(
   "default_rules",
   type: "success",
-  # TODO: When the notification period expires, uncomment the begin-end block to activate this test.
-  issues: :_,
-  #   issues: [
-  #     {
-  #       id: "runtime/printf",
-  #       path: "src/foo/sample.cpp",
-  #       location: { start_line: 42 },
-  #       message: 'Potential format string bug. Do printf("%s", str.c_str()) instead.',
-  #       links: [],
-  #       object: { confidence: "4" },
-  #       git_blame_info: {
-  #         commit: :_, line_hash: "2f1bd03f5ae88a98b3f874b1889c3ea965c58267", original_line: 42, final_line: 42
-  #       }
-  #     }
-  #   ],
-  # TODO: When the notification period expires, this warning message will no longer appear. Delete all warning entries in this file.
-  warnings: [{ message: <<~MSG.strip, file: "CPPLINT.cfg" }],
-    Sider's recommended configuration file is about to be released in mid October 2020.
-    After the release, Sider will automatically apply our recommended ruleset if you don't have the cpplint configuration file called `CPPLINT.cfg` in your repository.
-  MSG
+  issues: [
+    {
+      id: "runtime/printf",
+      path: "src/foo/sample.cpp",
+      location: { start_line: 43 },
+      message: 'Potential format string bug. Do printf("%s", str.c_str()) instead.',
+      links: [],
+      object: { confidence: "4" },
+      git_blame_info: {
+        commit: :_, line_hash: "2f1bd03f5ae88a98b3f874b1889c3ea965c58267", original_line: 43, final_line: 43
+      }
+    }
+  ],
   analyzer: { name: "cpplint", version: default_version }
 )
 
 s.add_test(
   "default_rules_when_cfg_exists",
   type: "success",
-  # TODO: When the notification period expires, uncomment the begin-end block to activate this test.
-  issues: :_,
-  #   issues: [
-  #     {
-  #       id: "legal/copyright",
-  #       path: "src/foo/test2.cpp",
-  #       location: nil,
-  #       message: 'No copyright message found.  You should have a line: "Copyright [year] <Copyright Owner>"',
-  #       links: [],
-  #       object: { confidence: "5" },
-  #       git_blame_info: nil,
-  #     },
-  #     {
-  #      id: "runtime/printf",
-  #      path: "src/foo/test2.cpp",
-  #      location: { start_line: 6 },
-  #      message: 'Potential format string bug. Do printf("%s", str) instead.',
-  #      links: [],
-  #      object: { confidence: "4" },
-  #      git_blame_info: {
-  #        commit: :_, line_hash: "de6ba426e696e32c9b26b0f6ac9f78413c365d91", original_line: 6, final_line: 6
-  #      }
-  #    }
-  #   ],
+  issues: [
+    {
+      id: "legal/copyright",
+      path: "src/foo/test2.cpp",
+      location: nil,
+      message: 'No copyright message found.  You should have a line: "Copyright [year] <Copyright Owner>"',
+      links: [],
+      object: { confidence: "5" },
+      git_blame_info: nil
+    },
+    {
+      id: "runtime/printf",
+      path: "src/foo/test2.cpp",
+      location: { start_line: 6 },
+      message: 'Potential format string bug. Do printf("%s", str) instead.',
+      links: [],
+      object: { confidence: "4" },
+      git_blame_info: {
+        commit: :_, line_hash: "de6ba426e696e32c9b26b0f6ac9f78413c365d91", original_line: 6, final_line: 6
+      }
+    }
+  ],
   analyzer: { name: "cpplint", version: default_version }
 )